feat(storage): move internal classes into Internal namespace and make them public
Refs: SITKO-CORE-T-4
