+++ src/Sitko.Core.Storage/Internal/StorageItemInfo.cs
@@ -1,0 +1,5 @@
+﻿using System;
+
+namespace Sitko.Core.Storage.Internal;
+
+public record StorageItemInfo(string Path, long FileSize, DateTimeOffset Date);
